pyproject.toml
[project]
name = "GenerativeKImitPython"
version = "1.0"
description = "Das ist unser Repository für den Kurs 'Generative KI mit Python'"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "chardet>=5.2.0",
    "chromadb>=0.5.20",
    "crewai[tools]>=0.83.0",
    "datasets>=3.1.0",
    "faiss-cpu>=1.9.0.post1",
    "haystack-ai>=2.6.1",
    "ipykernel>=6.29.5",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.3",
    "langchain-experimental>=0.3.3",
    "langchain-groq>=0.2.0",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.2.3",
    "langchain>=0.3.4",
    "langgraph>=0.2.39",
    "libmagic>=1.0",
    "matplotlib>=3.9.2",
    "pyarrow>=18.1.0",
    "python-dotenv>=1.0.1",
    "pytube>=15.0.0",
    "sentence-transformers>=3.3.0",
    "unstructured[pdf]>=0.16.8",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=0.6.3",
    "seaborn>=0.13.2",
    "ag2>=0.3.2",
    "langchain-anthropic>=0.3.10",
    "rich>=13.9.4",
    "markdown>=3.7",
    "pymupdf4llm>=0.0.18",
    "gensim>=4.3.3",
    "einops>=0.8.1",
    "create-mcp-server>=1.0.5",
    "sentencepiece>=0.2.0",
    "huggingface-hub>=0.26.2",
    "langchain-faiss>=0.1.1",
    "langchain-pymupdf4llm>=0.4.1",
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "nltk>=3.9.1",
    "streamlit>=1.50.0",
    "pytesseract>=0.3.13",
    "langchain-ollama>=0.3.0",
    "diffusers>=0.35.1",
    "kagglehub>=0.3.13",
    "python-docx>=1.2.0",
]

[tool.uv.sources]
swarm = { git = "ssh://git@github.com/openai/swarm.git" }

[tool.uv.workspace]
members = ["my-first-mcp-server"]
